move target from just the speaker to the center of the opening of speaker
